+++ codraft/tests/scenario_ima_app.py
@@ -25,6 +25,7 @@ from codraft.core.gui.processor.image import (
     PeakDetectionParam,
     ResizeParam,
     RotateParam,
+    WhiteTopHatParam,
     ZCalibrateParam,
 )
 from codraft.core.model.base import UniformRandomParam
@@ -73,6 +74,9 @@ def test_image_features(win: CodraFTMainWindow, data_size: int = 150) -> None:
 
     param = DenoiseWaveletParam()
     panel.processor.compute_denoise_wavelet(param)
+
+    param = WhiteTopHatParam()
+    panel.processor.compute_white_tophat(param)
 
     param = LogP1Param()
     param.n = 1