+++ CHANGELOG.md
@@ -11,6 +11,7 @@ New features:
   * Added Total variation denoising algorithm (TV Chambolle)
   * Added Bilateral filter denoising algorithm
   * Added Wavelet denoising algorithm
+  * Added white Top-Hat algorithm (disk footprint)
 * Added support for XYZ image files
 
 Bug fixes: